Create Code
This script uses the subprocess module to run the ping command, with the options -c 3 to send only 3 ping packets, and the host variable to specify the hostname or IP address of the network to ping. The stdout attribute of the result object is then decoded and returned.

You can call the function as ping('example.com') or ping('8.8.8.8')

Note that this script will only work on systems that have the ping command installed. On Windows, you need to use the ping command with the -n flag instead of -c.